another failed attempt at custom control.  templates are not loading
