+++ libwpfguts/RadioGroupBox.xaml
@@ -1,4 +1,4 @@
-<ItemsControl x:Class="libwpfguts.RadioGroupBox"
+<Control x:Class="libwpfguts.RadioGroupBox"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -6,38 +6,59 @@
              xmlns:l="clr-namespace:libwpfguts"
              mc:Ignorable="d" 
              d:DesignHeight="40" d:DesignWidth="100">
-    
-    <ItemsControl.Resources>
+
+    <Control.Resources>
         <Style TargetType="{x:Type RadioButton}">
             <Setter Property="Margin" Value="4,0,4,0" />
-            
         </Style>
-        
-    </ItemsControl.Resources>
-   
-    <ItemsControl.Template>
+    </Control.Resources>
+
+    <Control.Template>
         <ControlTemplate>
             <GroupBox Header="{Binding RelativeSource={RelativeSource AncestorType=l:RadioGroupBox}, Path=Header}">
-                <ItemsPresenter />
+                <ItemsControl x:Name="lstItems"
+                              ItemsSource="{Binding RelativeSource={RelativeSource AncestorType=l:RadioGroupBox}, Path=ItemsSource}">
+                    
+                    <ItemsPresenter />
+                    
+                    <ItemsControl.ItemsPanel>
+                        <ItemsPanelTemplate>
+                            <StackPanel Orientation="{Binding RelativeSource={RelativeSource AncestorType=l:RadioGroupBox}, Path=Orientation}" />
+                        </ItemsPanelTemplate>
+                    </ItemsControl.ItemsPanel>
+                    
+                    <ItemsControl.ItemTemplate>
+                        <DataTemplate>
+                            <RadioButton GroupName="{Binding RelativeSource={RelativeSource AncestorType=l:RadioGroupBox}, Path=Header}" 
+                                         Content="{Binding}">
+                            </RadioButton>
+                        </DataTemplate>
+                    </ItemsControl.ItemTemplate>
+                </ItemsControl>
             </GroupBox>
         </ControlTemplate>
-    </ItemsControl.Template>
-    
-    <ItemsControl.ItemsPanel>
-        <ItemsPanelTemplate>
-            <StackPanel Orientation="{Binding RelativeSource={RelativeSource AncestorType=l:RadioGroupBox}, Path=Orientation}" />
-        </ItemsPanelTemplate>
-    </ItemsControl.ItemsPanel>
+    </Control.Template>
+    <!--<ListBox ItemsSource="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=ItemsSource}"
+             SelectedItem="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=SelectedValue}">
+        <ListBox.Template>
+            <ControlTemplate>
+                
+            </ControlTemplate>
+        </ListBox.Template>
 
-    <ItemsControl.ItemTemplate>
-        <DataTemplate>
-            <RadioButton GroupName="{Binding RelativeSource={RelativeSource AncestorType=l:RadioGroupBox}, Path=Header}" 
-                         Content="{Binding Path=Item}"
-                         IsChecked="{Binding Path=IsSelected}">
-                <RadioButton.DataContext>
-                    <l:SelectedViewModel/>
-                </RadioButton.DataContext>
-            </RadioButton>
-        </DataTemplate>
-    </ItemsControl.ItemTemplate>
-</ItemsControl>
+        --><!--<ListBox.ItemsPanel>
+            <ItemsPanelTemplate>
+                <StackPanel Orientation="{Binding RelativeSource={RelativeSource AncestorType=l:RadioGroupBox}, Path=Orientation}" />
+            </ItemsPanelTemplate>
+        </ListBox.ItemsPanel>-->
+
+        <!--<ListBox.ItemTemplate>
+            <DataTemplate>
+                <RadioButton GroupName="{Binding RelativeSource={RelativeSource AncestorType=l:RadioGroupBox}, Path=Header}" 
+                             Content="{Binding}">
+                </RadioButton>
+            </DataTemplate>
+        </ListBox.ItemTemplate>--><!--
+    </ListBox>-->
+    
+</Control>
